change projected saving format to integer

+++ python_introduction/finance_calculator.py
@@ -3,7 +3,7 @@ total_monthly_expenses = int(input("Enter your total monthly expenses: "))
 
 monthly_savings = abs(total_monthly_expenses - monthly_income)
 
-projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
+projected_savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 print(f"Your monthly savings are ${monthly_savings}.")
 print(f"Projected savings after one year, with interest, is: ${projected_savings}")
